backend/requirements.txt
flask
flask-cors
opencv-python-headless
Pillow
numpy
scikit-learn
